Allow for item images to be updated
Added a conditional section to the item_form view which appears if the
item has a pre-existing image and the user is updating said item. The
section allows users to view and remove the items old image when
updating.

Updated the item_update_post controller function to allow for updating
item images.

Small style changes.